Restructure unit-test for stress eqlb.
